+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Overhaul of summary page
 - Overhaul of poco page
 - Moved services information to structure details window
+- Move reinforcement hour to structure details window
+- Add next reinforcement change info to structure details window
 - Combined ammo and fighter tabs in structure details window
 - Various performance improvements
 - Show owners on public poco page